animal-server/requirements.txt
flask
grpcio
grpcio-tools